pjax: PJAX Exclude Navigation
This addresses a performance issue where we reload navigation (which
includes queues, etc.) anytime a PJAX reload is called. (click New Ticket,
clicking PJAX refresh button, etc.) This is due to the main PJAX container
starting in the header, wrapping the entire navigation/content, and ending
in the footer. This commit adds a new PJAX sub-container to the
queue-tickets template which contains just the ticket queues. This also adds
a data attribute to the necessary links in the navigation and queue-tickets
template that points to the new PJAX subcontainer. Clicking any of the
ticket queues in navigation or sorting links will refresh only the queue
content and not the navigation which saves trips to the database increasing
performance slightly. (anything helps right) This will maintain current
functionality of full page to be reloaded if clicking on the Logo, clicking
the browser refresh button, or switching from Agent Panel to Admin Panel and
back.

In addition to all of the above, this adds a few lines of jQuery to
bold/unbold and add/remove Active classes on queues in the navigation. (to
mimic current "active queue" functionality)