issue: Inline Local Check
This addresses an issue where local inline image checks can be bypassed if
`cid:` is any where in the `src` url. This is due to the `strpos()` check
that only checks for `strpos($match[3], 'cid:') !== false`. This updates the
`strpos()` check to ensure the `cid:` attribute is in position `0` to ensure
local inline images.