mpdf: Unicode Spaces
This addresses an issue where Unicode Spaces like "hair spaces" (ie. U+200A)
cause fatal errors when attempting to print via mPDF. This is due to mPDF
not liking Unicode Spaces very much. This adds a check for Unicode Spaces
and if present will be replaced with a normal space.